storage-types: reorganize sources.rs into modules
This is a pure code movement PR that splits out
`src/storage-types/src/sources.rs` into the following structure:

```
sources
├── encoding.proto
├── encoding.rs
├── envelope.proto
├── envelope.rs
├── kafka.proto
├── kafka.rs
├── load_generator.proto
├── load_generator.rs
├── postgres.proto
├── postgres.rs
├── testscript.proto
└── testscript.rs
```
