+++ src/storage-types/build.rs
@@ -127,6 +127,11 @@ fn main() {
                 "storage-types/src/sinks.proto",
                 "storage-types/src/sources.proto",
                 "storage-types/src/sources/encoding.proto",
+                "storage-types/src/sources/envelope.proto",
+                "storage-types/src/sources/kafka.proto",
+                "storage-types/src/sources/postgres.proto",
+                "storage-types/src/sources/load_generator.proto",
+                "storage-types/src/sources/testscript.proto",
             ],
             &[".."],
         )
