+++ src/storage-types/src/sources/envelope.proto
@@ -1,0 +1,106 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+syntax = "proto3";
+
+import "google/protobuf/empty.proto";
+
+import "repr/src/global_id.proto";
+
+package mz_storage_types.sources.envelope;
+
+message ProtoSourceEnvelope {
+    oneof kind {
+        ProtoNoneEnvelope none = 1;
+        ProtoDebeziumEnvelope debezium = 2;
+        ProtoUpsertEnvelope upsert = 3;
+        google.protobuf.Empty cdc_v2 = 4;
+    }
+}
+
+message ProtoNoneEnvelope {
+    ProtoKeyEnvelope key_envelope = 1;
+    uint64 key_arity = 2;
+}
+
+message ProtoKeyEnvelope {
+    oneof kind {
+        google.protobuf.Empty none = 1;
+        google.protobuf.Empty flattened = 2;
+        string named = 3;
+    }
+}
+
+message ProtoUpsertEnvelope {
+    ProtoUpsertStyle style = 1;
+    repeated uint64 key_indices = 2;
+    uint64 source_arity = 3;
+    reserved 4;
+}
+
+message ProtoUpsertStyle {
+    message ProtoDebezium {
+        uint64 after_idx = 1;
+    }
+
+    oneof kind {
+        ProtoKeyEnvelope default = 1;
+        ProtoDebezium debezium = 2;
+    }
+}
+
+message ProtoDebeziumEnvelope {
+    uint64 before_idx = 1;
+    uint64 after_idx = 2;
+    ProtoDebeziumDedupProjection dedup = 3;
+}
+
+message ProtoDebeziumTransactionMetadata {
+    mz_repr.global_id.ProtoGlobalId tx_metadata_global_id = 1;
+    uint64 tx_status_idx = 2;
+    uint64 tx_transaction_id_idx = 3;
+    uint64 tx_data_collections_idx = 4;
+    uint64 tx_data_collections_data_collection_idx = 5;
+    uint64 tx_data_collections_event_count_idx = 6;
+    string tx_data_collection_name = 7;
+    uint64 data_transaction_idx = 8;
+    uint64 data_transaction_id_idx = 9;
+}
+
+message ProtoDebeziumDedupProjection {
+    uint64 op_idx = 1;
+    uint64 source_idx = 2;
+    uint64 snapshot_idx = 3;
+    ProtoDebeziumSourceProjection source_projection = 4;
+    ProtoDebeziumTransactionMetadata tx_metadata = 6;
+}
+
+message ProtoDebeziumSourceProjection {
+    message ProtoMySql {
+        uint64 file = 1;
+        uint64 pos = 2;
+        uint64 row = 3;
+    }
+
+    message ProtoPostgres {
+        uint64 sequence = 1;
+        uint64 lsn = 2;
+    }
+
+    message ProtoSqlServer {
+        uint64 change_lsn = 1;
+        uint64 event_serial_no = 2;
+    }
+
+    oneof kind {
+        ProtoMySql my_sql = 1;
+        ProtoPostgres postgres = 2;
+        ProtoSqlServer sql_server = 3;
+    }
+}
