+++ src/storage-types/src/sources/kafka.proto
@@ -1,0 +1,49 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+syntax = "proto3";
+
+import "google/protobuf/empty.proto";
+
+import "proto/src/proto.proto";
+import "repr/src/global_id.proto";
+import "storage-types/src/connections.proto";
+
+package mz_storage_types.sources.kafka;
+
+message ProtoKafkaSourceConnection {
+    reserved 5, 6, 7, 8, 9, 10, 12, 14;
+    mz_storage_types.connections.ProtoKafkaConnection connection = 1;
+    mz_repr.global_id.ProtoGlobalId connection_id = 13;
+    string topic = 2;
+    map<int32, int64> start_offsets = 3;
+    optional string group_id_prefix = 4;
+    repeated ProtoKafkaMetadataColumn metadata_columns = 11;
+    mz_proto.ProtoDuration topic_metadata_refresh_interval = 15;
+}
+
+message ProtoKafkaMetadataColumn {
+    string name = 1;
+    ProtoKafkaMetadataKind kind = 2;
+}
+
+message ProtoKafkaMetadataKind {
+    oneof kind {
+        google.protobuf.Empty partition = 1;
+        google.protobuf.Empty offset = 2;
+        google.protobuf.Empty timestamp = 3;
+        google.protobuf.Empty headers = 4;
+        ProtoKafkaHeader header = 5;
+    }
+}
+
+message ProtoKafkaHeader {
+    string key = 1;
+    bool use_bytes = 2;
+}
