+++ src/storage/src/source/generator/auction.rs
@@ -12,7 +12,8 @@ use std::iter;
 
 use mz_ore::now::{to_datetime, NowFn};
 use mz_repr::{Datum, Row};
-use mz_storage_types::sources::{Generator, MzOffset};
+use mz_storage_types::sources::load_generator::Generator;
+use mz_storage_types::sources::MzOffset;
 use rand::prelude::{Rng, SmallRng};
 use rand::seq::SliceRandom;
 use rand::SeedableRng;
