+++ src/storage/src/source/generator/datums.rs
@@ -11,7 +11,8 @@ use std::iter;
 
 use mz_ore::now::NowFn;
 use mz_repr::{Datum, Row, ScalarType};
-use mz_storage_types::sources::{Generator, MzOffset};
+use mz_storage_types::sources::load_generator::Generator;
+use mz_storage_types::sources::MzOffset;
 use timely::dataflow::operators::to_stream::Event;
 
 pub struct Datums {}
